% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsCCC.r
\name{ccc}
\alias{ccc}
\title{Create a Conditional Constant Correlation object}
\usage{
ccc(ntvgarchObj)
 .. or ..
ccc(nr.series)
}
\arguments{
\item{ntvgarchObj}{A valid ntvgarch object}

\item{nr.series}{Non-negative Integer. Will generate a simple CCC object (typically used for testing)}
}
\value{
A ccc_class object.  Suitable for use in Correlation Modelling.
}
\description{
\code{ccc} returns a CCC Object based on a specified number of series, or an ntvgarch object
}
\details{
I wanna write stuff
}
\note{
I am a note
}
