% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsCCC.r
\name{estimateCCC}
\alias{estimateCCC}
\title{Estimate a Conditional Constant Correlation object}
\usage{
estimateCCC(cccObj,estimationCtrl)
}
\arguments{
\item{cccObj}{A valid CCC object}

\item{estimationCtrl}{Optional. Provides controls for creating Standard Errors and displaying estimation progress}
}
\value{
A ccc_class object.
}
\description{
\code{estimateCCC} returns an Estimated CCC Object
}
\details{
A newly created ccc object represents a model specification.  When we estimate this model using:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  myCCC = estimateCCC(myCCC)
}\if{html}{\out{</div>}}

we will find all the estimated information has been added to the object, e.g. estimated parameters, log-likelihood value, etc.
This results in a single object that contains the model specification and the estimated values.
}
\note{
I am a note
}
