% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsTVGARCH.r
\name{estimateGARCH}
\alias{estimateGARCH}
\title{Estimates a garch model}
\usage{
estimateGARCH(e,garchObj,estimationControl)
}
\arguments{
\item{e}{An estimated TV model}

\item{garchObj}{Use Enum, garchtype$...}

\item{estimationControl}{A list}
}
\value{
A garch_class object.
}
\description{
\code{estimateGARCH} is a
}
\details{
This object

\if{html}{\out{<div class="sourceCode">}}\preformatted{  myGarch = estimateGARCH(e,myGarch,estimationControl)
}\if{html}{\out{</div>}}
}
\note{
I am a note
}
