% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsTVGARCH.r
\name{estimateGARCH_RollingWindow}
\alias{estimateGARCH_RollingWindow}
\title{Approximately estimates a garch model}
\usage{
estimateGARCH_RollingWindow(e,garchObj,estimationControl)
}
\arguments{
\item{e}{An estimated TV model}

\item{garchObj}{Use Enum, garchtype$...}

\item{estimationControl}{A list}
}
\value{
A garch_class object.
}
\description{
\code{estimateGARCH_RollingWindow} is a
}
\details{
This object

\if{html}{\out{<div class="sourceCode">}}\preformatted{  estimationControl$vartargetWindow = 400
  myGarch = estimateGARCH_RollingWindow(e,myGarch,estimationControl)
}\if{html}{\out{</div>}}
}
\note{
I am a note
}
