% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsTVGARCH.r
\name{estimateTV}
\alias{estimateTV}
\title{Estimates a tv model}
\usage{
estimateTV(e,tvObj,estimationControl)
}
\arguments{
\item{e}{An estimated TV model}

\item{tvObj}{Use Enum, garchtype$...}

\item{estimationControl}{A list}
}
\value{
A tv_class object.
}
\description{
\code{estimateTV} is a
}
\details{
This object

\if{html}{\out{<div class="sourceCode">}}\preformatted{  myTv = estimateTV(e,myTv,estimationControl)
}\if{html}{\out{</div>}}
}
\note{
I am a note
}
