% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsTVGARCH.r
\name{estimateTVGARCH}
\alias{estimateTVGARCH}
\title{Estimates a tv model}
\usage{
estimateTVGARCH(e,tvgarchObj,estimationControl)
}
\arguments{
\item{e}{An}

\item{tvgarchObj}{Use}

\item{estimationControl}{A list}
}
\value{
A tvgarch_class object.
}
\description{
\code{estimateTVGARCH} is a
}
\details{
This object

\if{html}{\out{<div class="sourceCode">}}\preformatted{  myTvGarch = estimateTVGARCH(e,myTvGarch,estimationControl)
}\if{html}{\out{</div>}}
}
\note{
I am a note
}
