% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsTVGARCH.r
\name{garch}
\alias{garch}
\title{A garch object}
\usage{
garch(type,order)
}
\arguments{
\item{type}{Use Enum, garchtype$...}

\item{order}{Optional. Integer, defaults to 1, e.g. Default is a Garch(1,1) model.}
}
\value{
A garch_class object.
}
\description{
\code{garch} is a
}
\details{
This object

\if{html}{\out{<div class="sourceCode">}}\preformatted{  myGarch = garch(garchtype$standard)
}\if{html}{\out{</div>}}

This package provide support for standard Garch and GJR-Garch specifications
}
\note{
I am a note
}
