% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsNTVGARCH.r
\name{ntvgarch}
\alias{ntvgarch}
\title{A list of estimated mtvgarch models}
\usage{
ntvgarch("tvgarch_list","series.names")
}
\arguments{
\item{tvgarch_list}{A standard R list() containing 2 or more univariate estimated mtvgarch models}

\item{series.names}{Optional. A character vector containing the names of the series in order.}
}
\value{
An ntvgarch_class object. Basically a named list with some useful attributes added.
}
\description{
\code{ntvgarch} is a convenience object designed to hold a list of estimated mtvgarch models, to be used for multivariate correlation modelling
}
\details{
This object was created to avoid issues with simple named lists.  By forcing the user to create this object, we can
ensure that all relevant validations are run and that the multivariate correlation models that use this will function as expected

\if{html}{\out{<div class="sourceCode">}}\preformatted{  myNTVG = ntvgarch("tvgarch_list","series.names")
}\if{html}{\out{</div>}}
}
\note{
I am a note
}
