% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsTVGARCH.r
\name{plot,tv_class,missing-method}
\alias{plot,tv_class,missing-method}
\title{plot.tv}
\usage{
plot(tvObj,...)
}
\arguments{
\item{...}{Other parameters that can be passed to the generic plot() function}

\item{tvObj}{Use Enum, garchtype$...}
}
\description{
\code{plot.tv} is a
}
\details{
This
}
\note{
I am a note
}
