% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsTVGARCH.r
\name{summary,tv_class-method}
\alias{summary,tv_class-method}
\title{summary.tv}
\usage{
summary(tvObj,...)
}
\arguments{
\item{...}{Other parameters that can be passed to the generic plot() function}

\item{tvObj}{Use Enum, garchtype$...}
}
\description{
\code{summary.tv} is a
}
\details{
This
}
\note{
I am a note
}
