% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsCCC.r
\name{test.CCCParsim}
\alias{test.CCCParsim}
\title{Parsimonious Test for Constancy in a CCC model}
\usage{
test.CCCParsim(H0,st,testOrder)
}
\arguments{
\item{H0}{Null Hypothesis of the test: A valid, estimated CCC object}

\item{st}{Numeric vector representing the smooth-transition variable}

\item{testOrder}{Integer = 1,2,3 or 4.}
}
\value{
A named 2x1 matrix, with the Test Statistic Value and the P Value.
}
\description{
\code{test.CCCParsim} uses Eigenvectors to reduce the numerical/processing complexity of the test
}
\details{
A newly created ccc object represents a model specification.  When we estimate this model using:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  testResult = test.CCCParsim(H0,st,testOrder)
}\if{html}{\out{</div>}}

we will find all the estimated information has been added to the object, e.g. estimated parameters, log-likelihood value, etc.
This results in a single object that contains the model specification and the estimated values.
}
\note{
I am a note
}
