% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsCCC.r
\name{test.CCCvSTCC1}
\alias{test.CCCvSTCC1}
\title{LM Test for Constancy in a CCC model}
\usage{
test.CCCvSTCC1(H0,st,testOrder)
}
\arguments{
\item{H0}{Null Hypothesis of the test: A valid, estimated CCC object}

\item{st}{Numeric vector representing the smooth-transition variable}

\item{testOrder}{Integer = 1,2,3 or 4.}
}
\value{
A named 2x1 matrix, with the Test Statistic Value and the P Value.
}
\description{
\code{test.CCCvSTCC1} uses black magic to determine if the correlations are constant over time
}
\details{
A newly created ccc object represents a model specification.  When we estimate this model using:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  testResult = test.CCCvSTCC1(H0,st,testOrder)
}\if{html}{\out{</div>}}
}
\note{
I am a note
}
