% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsTVGARCH.r
\name{test.LM.Robust}
\alias{test.LM.Robust}
\title{LM TR-Squared Test - Robust Version}
\usage{
test.LM.Robust(e,tvObj,testOrder)
}
\arguments{
\item{e}{An estimated TV model}

\item{tvObj}{Use Enum, garchtype$...}

\item{testOrder}{An integer}
}
\value{
A numeric test statistic
}
\description{
\code{test.LM.Robust} is a
}
\details{
This object

\if{html}{\out{<div class="sourceCode">}}\preformatted{ testOrder = 1
 testResult = test.LM.Robust(e,myTv,testOrder)
}\if{html}{\out{</div>}}
}
\note{
I am a note
}
