% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsTVGARCH.r
\name{test.misSpec1}
\alias{test.misSpec1}
\title{Tests for mis-specification of...}
\usage{
test.misSpec1(e,tvgarch,testOrd)
}
\arguments{
\item{e}{A numeric vector}

\item{tvgarch}{An estimated TvGarch model}

\item{testOrd}{Use Enum, garchtype$...}
}
\value{
A list...
}
\description{
\code{test.misSpec1} is a
}
\details{
This object
}
\note{
I am a note
}
