% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsTVGARCH.r
\name{test.misSpec4}
\alias{test.misSpec4}
\title{Tests for mis-specification of...}
\usage{
test.misSpec4(e,tvgarch,exogVar)
}
\arguments{
\item{e}{A numeric vector}

\item{exogVar}{Use Enum, garchtype$...}

\item{tvgarch}{An estimated TvGarch model}
}
\value{
A list...
}
\description{
\code{test.misSpec4} is a
}
\details{
This object
}
\note{
I am a note
}
