% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsTVGARCH.r
\name{testStatDist}
\alias{testStatDist}
\title{Generates a test statistic distribution...}
\usage{
testStatDist(refdata,tvObj,reftests,simcontrol)
}
\arguments{
\item{refdata}{A matrix}

\item{tvObj}{Use Enum, garchtype$...}

\item{reftests}{An integer}

\item{simcontrol}{An integer}
}
\value{
A list containing
}
\description{
\code{testStatDist} is a
}
\details{
This object

\if{html}{\out{<div class="sourceCode">}}\preformatted{ simcontrol$maxTestorder = 2
 testResult = testStatDist(refdata,tvObj,reftests,simcontrol)
}\if{html}{\out{</div>}}
}
\note{
I am a note
}
