% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsTVGARCH.r
\name{tv}
\alias{tv}
\title{A time-varying}
\usage{
tv(st,shape)
}
\arguments{
\item{st}{A smooth}

\item{shape}{Use enum, tvshape$...}
}
\value{
A tv_class object.
}
\description{
\code{tv} is a
}
\details{
This object

\if{html}{\out{<div class="sourceCode">}}\preformatted{  myTV = tv(st,tvshape$single)
}\if{html}{\out{</div>}}
}
\note{
I am a note
}
