% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsTVGARCH.r
\name{tvgarch}
\alias{tvgarch}
\title{A multiplicative tvgarch object}
\usage{
tvgarch(tvObj,garchType)
}
\arguments{
\item{tvObj}{An estimated TV model}

\item{garchType}{Use Enum, garchtype$...}
}
\value{
A tvgarch_class object.
}
\description{
\code{tvgarch} is a
}
\details{
This object

\if{html}{\out{<div class="sourceCode">}}\preformatted{  myTvGarch = tvgarch(myTV,garchtype$standard)
}\if{html}{\out{</div>}}
}
\note{
I am a note
}
